% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.linear.R
\name{confint.multinom.linear}
\alias{confint.multinom.linear}
\title{Compute a confidence interval by test inversion given p-values on a
grid of points in the parameter space}
\usage{
\method{confint}{multinom.linear}(ml, alpha = 0.05)
}
\arguments{
\item{ml}{an object of class multinom.linear}

\item{alpha}{level of the confidence interval}
}
\value{
a matrix with 2 columns, corresponding to the starting and
ending theta values that make up the confidence interval
}
\description{
Compute a confidence interval by test inversion given p-values on a
grid of points in the parameter space
}
\examples{
## see the documentation for multinom.linear for an example
}
