% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.linear.R
\name{p.sampler.1}
\alias{p.sampler.1}
\title{Sample points the intersection of the probability simpler with
$\\{c^tx=1\\}$}
\usage{
p.sampler.1(n, c, shape = 1)
}
\arguments{
\item{n}{the number of points to sample}

\item{c}{the coefficient vector}
}
\value{
a matrix of points in the intersection
}
\description{
This sampling routine is relatively fast but non-uniform. See the referenced manuscript for details.
}
