% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.linear.R
\name{p.sampler.2}
\alias{p.sampler.2}
\title{Sample points the intersection of the probability simpler with
$\\{c^tx=1\\}$}
\usage{
p.sampler.2(n, c, concentration = 1, tol = 1e-06)
}
\arguments{
\item{n}{the number of points to sample}

\item{c}{the coefficient vector}

\item{concentration}{dirichlet parameter; if NULL then
a deterministic sample is used}
}
\value{
a matrix of points in the intersection
}
\description{
This sampling routine is relatively slow but allows for
deterministic or stochastic sampling.  See the referenced
manuscript for details.
}
