% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.linear.R
\name{plot.multinom.linear}
\alias{plot.multinom.linear}
\title{Plot the p-values for the hypothesis test that is inverted to form
a confidence interval}
\usage{
\method{plot}{multinom.linear}(ml, alpha = 0.05, ...)
}
\arguments{
\item{ml}{an object of class multinom.linear}

\item{alpha}{level of the test; indicated in the plot by a horizontal line}

\item{...}{additional parameters passed to plot}
}
\value{
plot output
}
\description{
Plot the p-values for the hypothesis test that is inverted to form
a confidence interval
}
\examples{
## see the documentation for multinom.linear for an example

}
