% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.linear.R
\name{test.multinom.linear}
\alias{test.multinom.linear}
\title{Carry out a hypothesis test of the null that a particular value
theta generated the sample}
\usage{
test.multinom.linear(ml, theta.null)
}
\arguments{
\item{ml}{an object of class multinom.linear}

\item{theta.null}{the null value of the parameter to be tested}
}
\value{
an object of class htest excpt for "statistic" entry, since
multinom.linear doesnt currently expose the test statistics,
only the p-values
}
\description{
Carry out a hypothesis test of the null that a particular value
theta generated the sample
}
\examples{
## see the documentation for multinom.linear for an example

}
