% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.linear.R
\name{test.stat.1}
\alias{test.stat.1}
\title{Center and studentize observed multinomial proportions}
\usage{
test.stat.1(p.obs, p, c)
}
\arguments{
\item{p.obs}{the observed proportions}

\item{p}{the null proportions}

\item{c}{the coefficient vector applied to the proportions}
}
\value{
a vector of test statistics
}
\description{
Center and studentize observed multinomial proportions
}
