% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.linear.R
\name{test.stat.5}
\alias{test.stat.5}
\title{Center and studentize observed multinomial proportions with an
added regularization term}
\usage{
test.stat.5(p.obs, p, c, n, lambda = 0.1)
}
\arguments{
\item{p.obs}{the observed proportions}

\item{p}{the null proportions}

\item{c}{the coefficient vector applied to the proportions}
}
\value{
a vector of test statistics
}
\description{
Center and studentize observed multinomial proportions with an
added regularization term
}
